% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_vaa}
\alias{download_vaa}
\title{Download nhdplusVAA data from HydroShare}
\usage{
download_vaa(url = vaa_hydroshare)
}
\arguments{
\item{url}{Dont mess with this :)}
}
\value{
path to cached data
}
\description{
downloads and caches data on your computer
}
