% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_vaa}
\alias{get_vaa}
\title{Variable Subset}
\usage{
get_vaa(vars = NULL)
}
\arguments{
\item{vars}{The variable names you would like, always includes comid}
}
\value{
data.frame
}
\description{
Return NHDPlus Attributes
}
